+++ src/index.test.ts
@@ -27,6 +27,7 @@ import {
   createLitDataset,
   fetchLitDataset,
   unstable_fetchResourceInfoWithAcl,
+  isContainer,
   saveLitDatasetAt,
   saveLitDatasetInContainer,
   unstable_saveAclFor,
@@ -101,6 +102,9 @@ import {
   unstable_getAgentDefaultAccessModesOne,
   unstable_getAgentDefaultAccessModesAll,
   unstable_setAgentDefaultAccessModes,
+  unstable_getPublicAccessModes,
+  unstable_getPublicResourceAccessModes,
+  unstable_getPublicDefaultAccessModes,
   // Deprecated functions still exported for backwards compatibility:
   getStringUnlocalizedOne,
   getStringUnlocalizedAll,
@@ -124,6 +128,7 @@ it("exports the public API from the entry file", () => {
   expect(createLitDataset).toBeDefined();
   expect(fetchLitDataset).toBeDefined();
   expect(unstable_fetchResourceInfoWithAcl).toBeDefined();
+  expect(isContainer).toBeDefined();
   expect(saveLitDatasetAt).toBeDefined();
   expect(saveLitDatasetInContainer).toBeDefined();
   expect(unstable_saveAclFor).toBeDefined();
@@ -198,6 +203,9 @@ it("exports the public API from the entry file", () => {
   expect(unstable_getAgentDefaultAccessModesOne).toBeDefined();
   expect(unstable_getAgentDefaultAccessModesAll).toBeDefined();
   expect(unstable_setAgentDefaultAccessModes).toBeDefined();
+  expect(unstable_getPublicAccessModes).toBeDefined();
+  expect(unstable_getPublicResourceAccessModes).toBeDefined();
+  expect(unstable_getPublicDefaultAccessModes).toBeDefined();
 });
 
 it("still exports deprecated methods", () => {