+++ CHANGELOG.md
@@ -12,5 +12,5 @@ First release! What's possible with this first release:
 - Store data back to Solid Pods.
 - Read data from datasets.
 - Manipulate data in datasets.
-- Inspect a user's permissions w.r.t. a given Resource. (Experimental.)
-- Retrieve, delete and/or write any file (including non-RDF) from/to a Pod.
+- Inspect a user's and public permissions w.r.t. a given Resource. (Experimental.)
+- Retrieve, delete and/or write any file (including non-RDF) from/to a Pod. (Experimental.)